k8s: pod shell added
